backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
sqlalchemy==2.0.35
alembic==1.13.3
psycopg2-binary==2.9.9
pydantic==2.9.2
pydantic-settings==2.5.2
tweepy==4.14.0
apscheduler==3.10.4
cryptography==43.0.1
python-multipart==0.0.12
httpx==0.27.2
python-dotenv==1.0.1
Pillow==10.4.0
python-jose[cryptography]==3.3.0
bcrypt==4.2.1
email-validator==2.2.0
